refactor: create bin-inspector feature directory (Phase 5a)
Move Inspector components and useBinInspector hook to a new
feature-based directory structure:

- src/features/bin-inspector/components/Inspector/
  - SingleBinInspector, MultiBinInspector, EmptyState
  - SplitWarning, CustomPropertiesEditor
- src/features/bin-inspector/hooks/useBinInspector.ts
- Barrel exports at feature and component levels

Re-export stubs maintain backward compatibility at old locations
(src/components/Inspector/, src/hooks/useBinInspector.ts).
These will be removed in Phase 6.

Update test imports to use barrel exports.

Part of folder structure reorganization Phase 5.
