+++ src/components/Inspector/index.ts
@@ -1,9 +1,14 @@
-// Hook - re-exported from src/hooks for backwards compatibility
-export { useBinInspector } from '../../hooks/useBinInspector';
-export type { UseBinInspectorReturn, BinField, BinConstraints, ConfirmDeleteState } from '../../hooks/useBinInspector';
-
-// Components
-export { SplitWarning } from './SplitWarning';
-export { SingleBinInspector } from './SingleBinInspector';
-export { MultiBinInspector } from './MultiBinInspector';
-export { EmptyState } from './EmptyState';
+// Re-export from new location for backward compatibility
+// TODO: Remove in Phase 6 - update all imports to use features/bin-inspector
+export {
+  useBinInspector,
+  type BinField,
+  type BinConstraints,
+  type ConfirmDeleteState,
+  type UseBinInspectorReturn,
+  CustomPropertiesEditor,
+  EmptyState,
+  MultiBinInspector,
+  SingleBinInspector,
+  SplitWarning,
+} from '../../features/bin-inspector';
