+++ src/features/bin-inspector/components/Inspector/EmptyState.tsx
