+++ src/features/bin-inspector/components/Inspector/MultiBinInspector.tsx
@@ -1,9 +1,9 @@
 import { useState } from 'react';
-import { STAGING_ID, DEFAULT_CATEGORY_COLOR, CONSTRAINTS } from '../../core/constants';
+import { STAGING_ID, DEFAULT_CATEGORY_COLOR, CONSTRAINTS } from '../../../../core/constants';
 import type { UseBinInspectorReturn } from '../../hooks/useBinInspector';
-import type { Layer } from '../../core/types';
-import { SelectDropdown } from '../SelectDropdown';
-import { BulkIncrementControl } from '../BulkIncrementControl';
+import type { Layer } from '../../../../core/types';
+import { SelectDropdown } from '../../../../components/SelectDropdown';
+import { BulkIncrementControl } from '../../../../components/BulkIncrementControl';
 
 interface MultiBinInspectorProps {
   inspector: UseBinInspectorReturn;
