+++ src/features/bin-inspector/components/Inspector/SplitWarning.tsx
