+++ src/features/bin-inspector/components/Inspector/index.ts
@@ -1,0 +1,6 @@
+// Barrel export for Inspector components
+export { CustomPropertiesEditor } from './CustomPropertiesEditor';
+export { EmptyState } from './EmptyState';
+export { MultiBinInspector } from './MultiBinInspector';
+export { SingleBinInspector } from './SingleBinInspector';
+export { SplitWarning } from './SplitWarning';
