+++ src/features/bin-inspector/hooks/index.ts
@@ -1,0 +1,8 @@
+// Barrel export for bin-inspector hooks
+export { useBinInspector } from './useBinInspector';
+export type {
+  BinField,
+  BinConstraints,
+  ConfirmDeleteState,
+  UseBinInspectorReturn,
+} from './useBinInspector';
