+++ src/features/bin-inspector/index.ts
@@ -1,0 +1,3 @@
+// Barrel export for bin-inspector feature
+export * from './components/Inspector';
+export * from './hooks';
