+++ src/test/components/CustomPropertiesEditor.test.tsx
@@ -1,6 +1,6 @@
 import { describe, it, expect, vi, beforeEach } from 'vitest';
 import { render, fireEvent } from '@testing-library/react';
-import { CustomPropertiesEditor } from '../../components/Inspector/CustomPropertiesEditor';
+import { CustomPropertiesEditor } from '../../components/Inspector';
 import { CONSTRAINTS } from '../../core/constants';
 
 describe('CustomPropertiesEditor', () => {
