+++ src/test/components/MultiBinInspector.test.tsx
@@ -1,7 +1,7 @@
 import { describe, it, expect, vi, beforeEach } from 'vitest';
 import { render, screen, fireEvent } from '@testing-library/react';
-import { MultiBinInspector } from '../../components/Inspector/MultiBinInspector';
-import type { UseBinInspectorReturn } from '../../components/Inspector/useBinInspector';
+import { MultiBinInspector } from '../../components/Inspector';
+import type { UseBinInspectorReturn } from '../../components/Inspector';
 import { resetAllStores } from '../testUtils';
 
 describe('MultiBinInspector', () => {
