+++ src/test/components/SingleBinInspector.test.tsx
@@ -1,7 +1,7 @@
 import { describe, it, expect, vi, beforeEach } from 'vitest';
 import { render, screen, fireEvent } from '@testing-library/react';
-import { SingleBinInspector } from '../../components/Inspector/SingleBinInspector';
-import type { UseBinInspectorReturn } from '../../components/Inspector/useBinInspector';
+import { SingleBinInspector } from '../../components/Inspector';
+import type { UseBinInspectorReturn } from '../../components/Inspector';
 import { resetAllStores } from '../testUtils';
 import { useUIStore } from '../../core/store';
